Remove "vessel" property to avoid repeated calls to CurrentVessel()
SlingshotCore.FixedUpdate refers to "vessel" in a bunch of places, and
that was actually a property whose getter was a call to CurrentVessel().
That's a somewhat expensive operation, especially in the tracking
station (where it involves a call to FindObjectOfType(), which Unity
documents as "very slow" and not to be used every frame), so the
multiple "vessel" accesses made the UI very laggy when a vessel was
selected in the tracking station.

To avoid calling CurrentVessel() so often, I've removed the "vessel"
property and replaced it with a local variable in each method that needs
it, initialized with a single up-front call to CurrentVessel().  This
still isn't ideal (since it still calls FindObjectOfType() at least once
per FixedUpdate), but it's a simple change that provides a significant
performance improvement when a vessel is selected in the tracking
station.