Make SlingshotCore.CurrentVessel() an abstract method
Getting the current vessel is done differently depending on whether the
player is in the flight scene or tracking station, and CurrentVessel()
was a static method that would check HighLogic.LoadedScene to decide
what to do.  But there are already separate derived classes for the two
scenes, so the same thing can be accomplished with an abstract method
that's implemented differently in those two classes.  This avoids the
need to check HighLogic.LoadedScene when the calling object already
knows what scene it's in.