Cache the SpaceTracking lookup, since it's slow
Unity's FindObjectOfType method is slow, and calling it every UI frame
(from WindowGUI) has a major performance impact.  Since it's being used
to look up an object that's always the same for the lifetime of the
scene, we can do the lookup just once in Start() and save it in a field.

This resolves the laggy game UI when the Slingshotter window is open.